+++ tests/testthat/test-calc_indicator.R
@@ -1,7 +1,7 @@
 test_that("calc_indicator works", {
   aoi <- read_sf(
     system.file("extdata", "gfw_sample.gpkg",
-      package = "mapme.biodiversity"
+                package = "mapme.biodiversity"
     )
   )
 
@@ -58,22 +58,6 @@ test_that("calc_indicator works", {
     tolerance = 1e-3
   )
 
-  portfolio <- init_portfolio(
-    aoi,
-    years = 2000:2005,
-    outdir = outdir,
-    tmpdir = tmpdir,
-    add_resources = TRUE,
-    verbose = FALSE
-  )
-
-  stat <- calc_indicators(
-    portfolio,
-    indicators = "treecover_area",
-    min_size = 5,
-    min_cover = 30
-  )$treecover_area[[1]]
-
   expect_equal(
     names(stat),
     c("years", "treecover")
@@ -92,7 +76,7 @@ test_that("calc_indicator works", {
 test_that("Parallelization works", {
   aoi <- read_sf(
     system.file("extdata", "gfw_sample.gpkg",
-      package = "mapme.biodiversity"
+                package = "mapme.biodiversity"
     )
   )
 
@@ -146,7 +130,7 @@ test_that("Parallelization works", {
 
   expect_equal(
     stat,
-    c(2603.803, 2600.664, 2596.358, 2557.306, 2540.299, 2532.707),
+    c(2656.062, 2652.923, 2648.616, 2609.273, 2592.046, 2584.090),
     tolerance = 1e-3
   )
 })
@@ -216,7 +200,7 @@ test_that(".bind_assets works correctly", {
 test_that(".prep works correctly", {
   x <- read_sf(
     system.file("extdata", "gfw_sample.gpkg",
-      package = "mapme.biodiversity"
+                package = "mapme.biodiversity"
     )
   )
 
@@ -246,8 +230,8 @@ test_that(".prep works correctly", {
   )
 
   available_resources <- attr(x, "resources")
-  required_resources <- available_indicators("treecover_area")[[1]]$resources
-  output <- .prep(x, available_resources, required_resources)
+  required_resources <- available_indicators("treecover_area")[[1]][["resources"]]
+  output <- .prep_resources(x, available_resources, required_resources)
 
   expect_equal(
     length(output),
@@ -264,10 +248,15 @@ test_that(".prep works correctly", {
 
   x2 <- read_sf(list.files(
     system.file("extdata", package = "mapme.biodiversity"),
-    pattern = "shell_beach", full.names = TRUE
-  ))
-  required_resources <- available_indicators("mangroves_area")[[1]]$resources
-  output <- .prep(x2, available_resources, required_resources)
+    pattern = "shell_beach", full.names = TRUE )) %>%
+    init_portfolio(years = 2016,
+                     outdir = outdir,
+                     tmpdir = tmpdir,
+                     verbose = FALSE) %>%
+    get_resources("gmw")
+  available_resources <- attr(x2, "resources")
+  required_resources <- available_indicators("mangroves_area")[[1]][["resources"]]
+  output <- .prep_resources(x2, available_resources, required_resources)
 
   expect_equal(
     length(output),
@@ -291,7 +280,62 @@ test_that(".prep works correctly", {
   )
 
   expect_error(
-    .prep(x, available_resources, list(gmw = "sth")),
+    .prep_resources(x, available_resources, list(gmw = "sth")),
     "Resource type 'sth' currently not supported"
   )
+})
+
+
+test_that(".read_raster works correctly", {
+
+  dummy <- terra::rast()
+  dummy_splitted <- aggregate(dummy, fact = c(ceiling(nrow(dummy) / 4), ceiling(ncol(dummy) / 4)))
+  dummy_splitted[] <- 1:16
+  polys <- terra::as.polygons(dummy_splitted) %>% st_as_sf()
+  dummies <- lapply(1:nrow(polys), function(i) crop(dummy_splitted, polys[i, ]))
+  temp_loc <- tempfile()
+  dir.create(temp_loc, showWarnings = FALSE)
+  purrr::walk(1:length(dummies), function(i) {
+    writeRaster(dummies[[i]], filename = file.path(temp_loc, paste0("2000_tile_", i, ".tif")))
+    writeRaster(dummies[[i]], filename = file.path(temp_loc, paste0("2001_tile_", i, ".tif")))
+  })
+
+  files <- list.files(temp_loc, full.names = TRUE)
+  footprints <- .make_footprints(files)
+  x <- st_bbox(dummy) %>% st_as_sfc() %>% st_as_sf()
+  extent <- c(-180, 180, -90, 90)
+  names(extent) <- c("xmin", "xmax", "ymin", "ymax")
+
+  tiled_temporal <- .read_raster(x, footprints)
+  expect_equal(names(tiled_temporal), c("2000_tile_1", "2001_tile_1"))
+  expect_equal(as.vector(ext(tiled_temporal)), extent)
+
+  tiled <- .read_raster(x, footprints[grep("2001", footprints$location), ])
+  expect_equal(names(tiled), "2001_tile_1")
+  expect_equal(as.vector(ext(tiled)), extent)
+
+  temporal <- .read_raster(x, footprints[grep("tile_12.tif", footprints$location), ])
+  extent[c(1:4)] <- c(90, 180, -45, 0)
+  expect_equal(names(temporal), c("2000_tile_12", "2001_tile_12"))
+  expect_equal(as.vector(ext(temporal)), extent)
+
+  single <- .read_raster(x, footprints[grep("2000_tile_10.tif", footprints$location), ])
+  extent[c(1:4)] <- c(-90, 0, -45, 0)
+  expect_equal(names(single), "2000_tile_10")
+  expect_equal(as.vector(ext(single)), extent)
+
+  expect_error(.read_raster(x, footprints[1:24, ]))
+
+})
+
+
+test_that(".check_single_asset works correctly", {
+  expect_warning(out <- .check_single_asset(NA, 1))
+  expect_equal(out, NA)
+  expect_warning(out <- .check_single_asset(try("a" + 1, silent = TRUE), 1))
+  expect_equal(out, NA)
+  expect_warning(out <- .check_single_asset(c(1:10), 1))
+  expect_equal(out, NA)
+  expect_warning(out <- .check_single_asset(tibble(), 1))
+  expect_equal(out, NA)
 })