+++ tests/testthat/test-get_teow.R
@@ -18,13 +18,13 @@ test_that(".get_teow works", {
     years = 2000:2010,
     outdir = outdir,
     tmpdir = tmpdir,
-    add_resources = FALSE,
     verbose = FALSE
   )
   # Add testing attribute in order to skip downloads
   attributes(portfolio)$testing <- TRUE
-  expect_snapshot(
-    .get_teow(portfolio) %>%
-      basename()
+  expect_equal(
+    .get_teow(portfolio, rundir = file.path(outdir, "teow")) %>%
+      basename(),
+    "wwf_terr_ecos.gpkg"
   )
 })